Add documentation; Simplify PoolingImageLoader
