+++ UpdateDB/DataLoading/ItemImageLoader.cs
@@ -18,6 +18,7 @@ namespace UpdateDB.DataLoading
 
         private static readonly ILog Log = LogManager.GetLogger(typeof(ItemImageLoader));
 
+        // the wiki's item classes for which images are retrieved
         private static readonly IReadOnlyList<string> RelevantWikiClasses = new[]
         {
             "One Hand Axes", "Two Hand Axes", "Bows", "Claws", "Daggers",
@@ -29,6 +30,7 @@ namespace UpdateDB.DataLoading
 
         private readonly bool _overwriteExisting;
 
+        /// <param name="overwriteExisting">if false, only images whose files don't exist are downloaded</param>
         public ItemImageLoader(bool overwriteExisting)
         {
             _overwriteExisting = overwriteExisting;
@@ -45,19 +47,22 @@ namespace UpdateDB.DataLoading
 
         private async Task ReadJson(string wikiClass)
         {
+            // for items with Normal rarity that can drop and have the given class ...
             var conditions = new ConditionBuilder
             {
                 {RdfRarity, "Normal"},
                 {RdfDropEnabled, "true"},
                 {RdfItemClass, wikiClass}
             };
+            // ... retrieve name and the icon page
             var printouts = new[] {RdfName, RdfIcon};
-
-            var results = (from ps in await WikiApiAccessor.AskArgs(conditions, printouts)
-                           let title = ps[RdfIcon].First.Value<string>("fulltext")
-                           let name = SingularValue<string>(ps, RdfName)
+            // the icon page entry is an array, the page title is the "fulltext" property in its first element
+            var results = (from result in await WikiApiAccessor.AskArgs(conditions, printouts)
+                           let title = result[RdfIcon].First.Value<string>("fulltext")
+                           let name = SingularValue<string>(result, RdfName)
                            select new {name, title}).ToList();
 
+            // query the image url in each icon page and save them
             var titleToName = results.ToDictionary(x => x.title, x => x.name);
             foreach (var tuple in await WikiApiAccessor.QueryImageInfoUrls(results.Select(t => t.title)))
             {
