+++ WPFSKillTree/Model/Items/ItemImage.cs
@@ -6,6 +6,10 @@ using POESKillTree.Utils;
 
 namespace POESKillTree.Model.Items
 {
+    /// <summary>
+    /// Represents an asynchronously loaded image for an item group or item base or from an url stored in an item's
+    /// json object.
+    /// </summary>
     public class ItemImage : Notifier
     {
         private static readonly ILog Log = LogManager.GetLogger(typeof(ItemImage));
@@ -15,12 +19,18 @@ namespace POESKillTree.Model.Items
         private readonly ItemImageService _itemImageService;
 
         private NotifyingTask<ImageSource> _imageSource;
+        /// <summary>
+        /// Gets a <see cref="NotifyingTask{TResult}"/> for the represented image.
+        /// </summary>
         public NotifyingTask<ImageSource> ImageSource
         {
             get { return _imageSource; }
             private set { SetProperty(ref _imageSource, value); }
         }
 
+        /// <summary>
+        /// Represents an image for an item group. The image will be loaded synchronously.
+        /// </summary>
         public ItemImage(ItemImageService itemImageService, ItemGroup baseGroup)
         {
             _itemImageService = itemImageService;
@@ -32,6 +42,10 @@ namespace POESKillTree.Model.Items
             );
         }
 
+        /// <summary>
+        /// Represents an image for an item base. First the group's image will be loaded synchronously,
+        /// which is then used as the image until the base item's image is loaded asynchronously.
+        /// </summary>
         public ItemImage(ItemImageService itemImageService, string baseName, ItemGroup baseGroup)
         {
             _itemImageService = itemImageService;
@@ -43,6 +57,10 @@ namespace POESKillTree.Model.Items
             );
         }
 
+        /// <summary>
+        /// Represents an image that is loaded asynchronously from an url. Only urls stored in an item's json as
+        /// retrieved from the official api are supported.
+        /// </summary>
         public ItemImage(ItemImage baseItemImage, string imageUrl)
         {
             _itemImageService = baseItemImage._itemImageService;
@@ -56,7 +74,6 @@ namespace POESKillTree.Model.Items
 
         private void NewImageSourceTask(Task<ImageSource> task, string errorMessage, ImageSource defaultValue)
         {
-            // todo use AsyncEx NotifyTaskCompletion instead
             ImageSource = new NotifyingTask<ImageSource>(task, e => Log.Error(errorMessage, e))
             {
                 Default = defaultValue
@@ -65,6 +82,7 @@ namespace POESKillTree.Model.Items
 
         private static string MakeUrl(string imageUrl)
         {
+            // if the image's url has no domain, the domain is the official site
             if (imageUrl.StartsWith("/"))
             {
                 return OfficialSiteUrl + imageUrl;
