+++ WPFSKillTree/Utils/WikiApi/WikiApiUtils.cs
@@ -14,22 +14,38 @@ namespace POESKillTree.Utils.WikiApi
         /// </summary>
         private const double ItemImageResizeFactor = 0.6;
 
+        /// <summary>
+        /// Returns the singular value of the predicate in the printouts.
+        /// </summary>
         public static T SingularValue<T>(JToken printouts, string rdfPredicate)
         {
             return printouts[rdfPredicate].First.Value<T>();
         }
 
+        /// <summary>
+        /// Returns the singular value of the predicate in the printouts or <paramref name="defaultValue"/> if the
+        /// predicate is not in the printouts.
+        /// </summary>
         public static T SingularValue<T>(JToken printouts, string rdfPredicate, T defaultValue)
         {
             var token = printouts[rdfPredicate];
             return token.HasValues ? token.First.Value<T>() : defaultValue;
         }
 
+        /// <summary>
+        /// Returns the plural value of the predicate in the printouts.
+        /// </summary>
         public static IEnumerable<T> PluralValue<T>(JToken printouts, string rdfPredicate)
         {
             return printouts[rdfPredicate].Values<T>();
         }
 
+        /// <summary>
+        /// Saves the image to a stream.
+        /// </summary>
+        /// <param name="imageData">the binary image data</param>
+        /// <param name="outputStream">the stream to save the image to</param>
+        /// <param name="resize">true iff the image is from the wiki and should be resized to match the stash</param>
         public static void SaveImage(byte[] imageData, Stream outputStream, bool resize)
         {
             using (var ms = new MemoryStream(imageData))
