+++ WPFSKillTree/WPFSKillTree.csproj
@@ -208,9 +208,8 @@
     <Compile Include="Model\Items\ItemImageService.cs" />
     <Compile Include="Utils\Extensions\NumberExtensions.cs" />
     <Compile Include="Utils\WikiApi\ApiAccessor.cs" />
-    <Compile Include="Utils\WikiApi\ApiConditionBuilder.cs" />
+    <Compile Include="Utils\WikiApi\ConditionBuilder.cs" />
     <Compile Include="Utils\WikiApi\PoolingImageLoader.cs" />
-    <Compile Include="Utils\WikiApi\PropertyBuilder.cs" />
     <Compile Include="Utils\WikiApi\ItemRdfPredicates.cs" />
     <Compile Include="Utils\WikiApi\WikiApiUtils.cs" />
     <Compile Include="Utils\Wpf\ComboBoxTemplateSelector.cs" />
